Force angularjs version for now
In later versions of angular, their jqlite code modifies the Node
prototype. In the version of jsdom (included in benv) I have to
use, Node is not defined. I can't upgrade to later versions of
benv or jsdom because they require Node >4 and I would like to
test the server side components on older versions of Node, at
least until I understand the differences better.

We are really very minimally using angularjs / jsdom anyway so
it's probably not that big of a deal to use an older version.